split views into folder structure
